move subpixel position tests out of the image subfolder
moved to layers/_tests as these tests involve multiple layer types
