+++ napari/layers/_tests/test_layer_attributes.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 
 from napari._tests.utils import layer_test_data
-from napari.layers import Image
+from napari.layers import Image, Labels
 
 
 @pytest.mark.parametrize(
@@ -54,3 +54,72 @@ def test_update_data_updates_layer_extent_cache(Layer, data, ndim):
     new_extent = layer.extent
     assert old_extent is not layer.extent
     assert new_extent is layer.extent
+
+
+def _check_subpixel_values(layer, val_dict):
+    ndisplay = layer._ndisplay
+    for center, expected_value in val_dict.items():
+        # ensure all positions within the pixel extent report the same value
+        # note: world=False so do not multiply offsets by the layer scale
+        for offset_0 in [-0.4999, 0, 0.4999]:
+            for offset_1 in [-0.4999, 0, 0.4999]:
+                position = [center[0] + offset_0, center[1] + offset_1]
+                view_direction = None
+                dims_displayed = None
+                if ndisplay == 3:
+                    position = [0] + position
+                    if isinstance(layer, Labels):
+                        # Labels implements _get_value_3d, Image does not
+                        view_direction = np.asarray([1.0, 0, 0])
+                        dims_displayed = [0, 1, 2]
+
+                val = layer.get_value(
+                    position=position,
+                    view_direction=view_direction,
+                    dims_displayed=dims_displayed,
+                    world=False,
+                )
+                assert val == expected_value
+
+
+@pytest.mark.parametrize('ImageClass', [Image, Labels])
+@pytest.mark.parametrize('ndim', [2, 3])
+def test_get_value_at_subpixel_offsets(ImageClass, ndim):
+    """check value at various shifts within a pixel/voxel's extent"""
+    if ndim == 3:
+        data = np.arange(1, 9).reshape(2, 2, 2)
+    elif ndim == 2:
+        data = np.arange(1, 5).reshape(2, 2)
+
+    # test using non-uniform scale per-axis
+    layer = ImageClass(data, scale=(0.5, 1, 2)[:ndim])
+    layer._slice_dims([0] * ndim, ndisplay=ndim)
+
+    # dictionary of expected values at each voxel center coordinate
+    val_dict = {
+        (0, 0): data[(0,) * (ndim - 2) + (0, 0)],
+        (0, 1): data[(0,) * (ndim - 2) + (0, 1)],
+        (1, 0): data[(0,) * (ndim - 2) + (1, 0)],
+        (1, 1): data[(0,) * (ndim - 2) + (1, 1)],
+    }
+    _check_subpixel_values(layer, val_dict)
+
+
+@pytest.mark.parametrize('ImageClass', [Image, Labels])
+def test_get_value_3d_view_of_2d_image(ImageClass):
+    """check value at various shifts within a pixel/voxel's extent"""
+    data = np.arange(1, 5).reshape(2, 2)
+
+    ndisplay = 3
+    # test using non-uniform scale per-axis
+    layer = ImageClass(data, scale=(0.5, 1))
+    layer._slice_dims([0] * ndisplay, ndisplay=ndisplay)
+
+    # dictionary of expected values at each voxel center coordinate
+    val_dict = {
+        (0, 0): data[(0, 0)],
+        (0, 1): data[(0, 1)],
+        (1, 0): data[(1, 0)],
+        (1, 1): data[(1, 1)],
+    }
+    _check_subpixel_values(layer, val_dict)
